Use a consistent configuration file name
Before, the initial configuration was explicitly drawn from "kdedrc",
while autoloading stuff was drawn from the application-default file,
which was "kded5rc".  This change makes it always use the
application-default file, which involves rearrangeing some of the code
to make sure the application instance always has its name set
consistently.

This also changes the configuration entry reading so that the defaults
are set in exactly one place.

REVIEW: 116991