Install .desktop file for kded5
Summary:
This is useful as on wayland setting a path to a desktop file is
mandatory for icons on any windows. Qt sets the app_id to org.kde.kde5
but this desktop file doesn't exist.

We do a similar fix in plasmashell and other services that don't really
map to the traditional application pattern.

I set the type to something other than Application so we don't have to
put an Exec line though it wouldn't really matter as it's hidden anway.

BUG: 387556

Test Plan:
Opened a vault
Got an icon

Reviewers: ivan

Reviewed By: ivan

Subscribers: kde-frameworks-devel

Tags: #frameworks

Differential Revision: https://phabricator.kde.org/D19986