Drop delayed second phase
kded runs Phase 0 and Phase 1 on startup and then plasma-session Startup calls Phase 2 somewhere around the end of startup.
Investigation and testing shows this has little effect on startup, if any, makes it slower.
More importantly, it complicates startup and makes it harder to recreate the startup sequence as systemd units.

Differential Revision: https://phabricator.kde.org/D28919