Shorten KSycoca update delay
Currently kded waits 10 seconds before KSycoca is updated.
This is a very long and user noticeable time, for example on a
fresh system without a browser after installing one it will take
that long after the process is finished until the icon appears in
the default Plasma panel.
Historically the timeout was changed in in the before times in
both directions, see kdelibs commits
https://commits.kde.org/unmaintained/kdelibs/f8b972cbc797dca46e63072248cee0daec78f46f
https://commits.kde.org/unmaintained/kdelibs/cc93a95616c8f4d24941dc2ebe1837327153dc3b
https://commits.kde.org/unmaintained/kdelibs/f6ae8d6490f5b0cf18b92337405971018ff568aa
It seems the intention is to batch events together that come in bursts
for example during a system update. A shorter timer will still
achieve the same goal, the wait period starts anew on every change
and subsequent call to QTimer::start. The difference is the delay
between when everything has calmed down to when KSyCoca is updated.