Watch for icons directory changes from KDED
Upon KIconTheme construction folders are checked to exist and only
existing ones are used later for lookup of icons.
Now consider that  a new program is installed and it installs its
icon into a  directory inside XDG_DATA_DIRS that before did either not
contain an 'icons' directory at all or there was no folder of
the specific theme  at this specific location.
A typically long running program that  wants to display such icons
like an  application launcher, taskmanager, system monitor  or
window manager won't be able to find these icons.
KDED can watch for 'icons' directories or new theme directories
(subfolders of 'icons' and notify running applications.
KDED seems the correct place to watch for such events, it alreay
does so for a bunch of subfolders of XDG_DATA_DIRS.
BUG:455702
FIXED-IN:5.101