Use kded5rc as config file name
KSharedConfig::openConfig() uses a file name based on the application name, in our case kded5rc.

The moment we rename to kded6 this changes the implicit config file name to kded6rc. This would cause existing
config in kded5rc to be ignored. To avoid that explicitly use kded5rc even when we are kded6